% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_fd.R
\name{find_fd}
\alias{find_fd}
\title{Find unconditional FDs in a directed acyclic graph (DAG)}
\usage{
find_fd(dag, X, Y, verbose = TRUE)
}
\arguments{
\item{dag}{DAG (in dagitty) in which front-doors should be found.}

\item{X}{Exposure (String).}

\item{Y}{Outcome (String).}

\item{verbose}{Whether to print intermediary output (boolean).}
}
\value{
A list of valid front-door mediator sets (each element is a
character vector of node names). Returns \code{NULL} invisibly if no
valid front-door exists.
}
\description{
Find unconditional FDs in a directed acyclic graph (DAG)
}
\examples{
dag1 <- dagitty::dagitty("dag {
X -> M
M -> Y
X [pos=\"0,0\"]
  M [pos=\"1,0\"]
  Y [pos=\"2,0\"]
}")

find_fd(dag1, "X", "Y")

dag2 <- dagitty::dagitty("dag {
A -> B
B -> C
U1 -> A
U1 -> B
U2 -> B
U2 -> C
U1 [pos=\"0,1\"]
  A  [pos=\"0,0\"]
  B  [pos=\"1,0.5\"]
  C  [pos=\"2,0\"]
  U2 [pos=\"2,1\"]
}")

find_fd(dag2, "A", "C", verbose=FALSE)
}
